Fixes post hoc analysis bug for running single variate after double variate.
Previously if you ran an analysis with two factors through to post hoc and then
went back and changed to a single factor analysis you'd get an error in the
post hoc run because the second independent variable was still lurking from the
previous run because the input$independent.variable.two doesn't ever get set
back to NULL because there is no longer a second selectInput. I tried all
kinds of things to fix this like:

- use updateSelectInput: not sure why this wouldn't work, maybe because the
  second selector didn't exist by the time this was run?
- used shinyjs::reset() but this doesn't work with ui elements generated on the
  server side with renderUI()
- used shinyjs::hidden() to hide the second selector. I think hidden() causes
  the selector to not function and so it doesn't set the value of
  input$indpendent.variable.

This took about 3 hours and it ended up being a simple solution.